Switch new and legacy aggregation chart url mapping
