+++ frontend/src/app/modules/result-selection/models/chart-switch-menu-entry.model.ts
@@ -2,4 +2,5 @@ export class ChartSwitchMenuEntry {
   baseUrl: string;
   label: string;
   icon: string;
+  devUrl?: string;
 }
