+++ frontend/src/app/modules/shared/components/chart-switch-menu/chart-switch-menu.component.ts
@@ -10,10 +10,10 @@ import {ChartSwitchMenuEntry} from "../../../result-selection/models/chart-switc
 export class ChartSwitchMenuComponent implements OnInit {
 
   chartSwitchMenu: ChartSwitchMenuEntry[] = [
-    {baseUrl: "/eventResultDashboardDev/showAll", label: "frontend.de.iteratec.osm.results.timeSeries", icon: "fas fa-chart-line"},
-    {baseUrl: "/aggregationDev/show", label: "frontend.de.iteratec.osm.results.aggregation", icon: "fas fa-chart-bar"},
-    {baseUrl: "/distributionDev/show", label: "frontend.de.iteratec.osm.results.distribution", icon: "fas fa-chart-area"},
-    {baseUrl: "/pageComparison/show", label: "frontend.de.iteratec.osm.results.pageComparison", icon: "fas fa-balance-scale"},
+    {baseUrl: "/eventResultDashboard/showAll", label: "frontend.de.iteratec.osm.results.timeSeries", icon: "fas fa-chart-line", devUrl: "/eventResultDashboardDev/showAll"},
+    {baseUrl: "/aggregation/show", label: "frontend.de.iteratec.osm.results.aggregation", icon: "fas fa-chart-bar"},
+    {baseUrl: "/distributionChart/show", label: "frontend.de.iteratec.osm.results.distribution", icon: "fas fa-chart-area", devUrl: "/distributionDev/show"},
+    {baseUrl: "/pageComparison/show", label: "frontend.de.iteratec.osm.results.pageComparison", icon: "fas fa-balance-scale", },
     {baseUrl: "/detailAnalysis/show", label: "frontend.de.iteratec.osm.results.detailAnalysis", icon: "fas fa-chart-pie"},
     {baseUrl: "/tabularResultPresentation/listResults", label: "frontend.de.iteratec.osm.results.resultList", icon: "fas fa-th-list"}
   ];
