+++ grails-app/controllers/de/iteratec/osm/result/AggregationController.groovy
@@ -1,6 +1,5 @@
 package de.iteratec.osm.result
 
-import de.iteratec.osm.OsmConfigCacheService
 import de.iteratec.osm.annotations.RestAction
 import de.iteratec.osm.barchart.BarchartAggregation
 import de.iteratec.osm.barchart.BarchartAggregationService
@@ -9,7 +8,6 @@ import de.iteratec.osm.csi.Page
 import de.iteratec.osm.measurement.schedule.Job
 import de.iteratec.osm.measurement.schedule.JobDaoService
 import de.iteratec.osm.measurement.schedule.JobGroup
-import de.iteratec.osm.measurement.schedule.JobGroupService
 import de.iteratec.osm.measurement.script.PlaceholdersUtility
 import de.iteratec.osm.measurement.script.Script
 import de.iteratec.osm.measurement.script.ScriptParser
@@ -18,58 +16,15 @@ import de.iteratec.osm.result.dto.AggregationChartSeriesDTO
 import de.iteratec.osm.util.ControllerUtils
 import de.iteratec.osm.util.ExceptionHandlerController
 import de.iteratec.osm.util.I18nService
-import de.iteratec.osm.util.PerformanceLoggingService
 import org.springframework.http.HttpStatus
+
 
 class AggregationController extends ExceptionHandlerController {
 
-    public final static String DATE_FORMAT_STRING_FOR_HIGH_CHART = 'dd.mm.yyyy';
-    public final static int MONDAY_WEEKSTART = 1
-
-    JobGroupService jobGroupService
     JobDaoService jobDaoService
-    EventResultDashboardService eventResultDashboardService
     I18nService i18nService
     PageService pageService
-    OsmConfigCacheService osmConfigCacheService
     BarchartAggregationService barchartAggregationService
-    PerformanceLoggingService performanceLoggingService
-
-    def index() {
-        redirect(action: 'show')
-    }
-
-    def show() {
-        Map<String, Object> modelToRender = [:]
-
-        // AggregatorTypes
-        modelToRender.put('aggrGroupValuesUnCached', SelectedMeasurand.createDataMapForOptGroupSelect())
-
-        // JobGroups
-        List<JobGroup> jobGroups = eventResultDashboardService.getAllJobGroups()
-        modelToRender.put('folders', jobGroups)
-        modelToRender.put('selectedFolder', params.selectedFolder)
-
-        // Pages
-        List<Page> pages = eventResultDashboardService.getAllPages()
-        modelToRender.put('pages', pages)
-        modelToRender.put('selectedPages', params.selectedPages)
-
-        // JavaScript-Utility-Stuff:
-        modelToRender.put("dateFormat", DATE_FORMAT_STRING_FOR_HIGH_CHART)
-        modelToRender.put("weekStart", MONDAY_WEEKSTART)
-
-        modelToRender.put('selectedAggrGroupValuesUnCached', [])
-
-        modelToRender.put("tagToJobGroupNameMap", jobGroupService.getTagToJobGroupNameMap())
-
-        // Done! :)
-        return modelToRender
-    }
-
-    def entryAndFollow() {
-        return [:]
-    }
 
     /**
      * Rest Method for ajax call.
@@ -117,9 +72,9 @@ class AggregationController extends ExceptionHandlerController {
 
         if (!chartDto.series) {
             ControllerUtils.sendSimpleResponseAsStream(
-                response,
-                HttpStatus.OK,
-                i18nService.msg("de.iteratec.ism.no.data.on.current.selection.heading", "No data")
+                    response,
+                    HttpStatus.OK,
+                    i18nService.msg("de.iteratec.ism.no.data.on.current.selection.heading", "No data")
             )
         } else {
             ControllerUtils.sendObjectAsJSON(response, chartDto)
@@ -148,18 +103,18 @@ class AggregationController extends ExceptionHandlerController {
                 return []
             } else {
                 return [new AggregationChartSeriesDTO(
-                    unit: it.unit,
-                    measurandLabel: it.measurandLabel,
-                    measurand: it.measurandName,
-                    measurandGroup: it.measurandGroup,
-                    value: it.value,
-                    valueComparative: it.valueComparative,
-                    page: it.page?.name,
-                    jobGroup: it.jobGroup.name,
-                    browser: it.browser?.name,
-                    aggregationValue: it.aggregationValue,
-                    deviceType: it?.deviceType,
-                    operatingSystem:  it?.operatingSystem
+                        unit: it.unit,
+                        measurandLabel: it.measurandLabel,
+                        measurand: it.measurandName,
+                        measurandGroup: it.measurandGroup,
+                        value: it.value,
+                        valueComparative: it.valueComparative,
+                        page: it.page?.name,
+                        jobGroup: it.jobGroup.name,
+                        browser: it.browser?.name,
+                        aggregationValue: it.aggregationValue,
+                        deviceType: it?.deviceType,
+                        operatingSystem:  it?.operatingSystem
                 )]
             }
         }
