+++ grails-app/views/aggregationLegacy/_adjustBarchartModal.gsp
