+++ src/test/groovy/de/iteratec/osm/result/AggregationControllerSpec.groovy
@@ -1,147 +1,22 @@
 package de.iteratec.osm.result
 
-import de.iteratec.osm.barchart.BarchartAggregation
-import de.iteratec.osm.csi.CsiConfiguration
-import de.iteratec.osm.csi.Page
-import de.iteratec.osm.measurement.schedule.JobGroup
-import de.iteratec.osm.result.dto.AggregationChartSeriesDTO
-import de.iteratec.osm.util.I18nService
-import grails.buildtestdata.BuildDomainTest
-import grails.buildtestdata.mixin.Build
-import grails.testing.web.controllers.ControllerUnitTest
+import grails.test.mixin.TestFor
 import spock.lang.Specification
 
-@Build([Page, JobGroup, CsiConfiguration])
-class AggregationControllerSpec extends Specification implements BuildDomainTest<Page>,
-        ControllerUnitTest<AggregationController> {
-
-    private AggregationController controllerUnderTest
-
-    private Page page1
-    private Page page2
-    private Page page3
+/**
+ * See the API for {@link grails.test.mixin.web.ControllerUnitTestMixin} for usage instructions
+ */
+@TestFor(AggregationController)
+class AggregationControllerSpec extends Specification {
 
     def setup() {
-        controllerUnderTest = controller
-
-        page1 = Page.build(name: "page one")
-        page2 = Page.build(name: "page two")
-        page3 = Page.build(name: "page three")
     }
 
-    def "test merge two lists that are equal"() {
-        given: "two lists that are equal"
-        List<Page> list1 = [page1, page2]
-        List<Page> list2 = [page1, page2]
-
-        when: "merging lists"
-        List<Page> result = controllerUnderTest.getOrderedPagesOfAllScripts([list1, list2])
-
-        then: "result should be equal too"
-        result == list1
+    def cleanup() {
     }
 
-    def "test merge two different lists with same length"() {
-        given: "two lists that are not equal"
-        List<Page> list1 = [page1, page2]
-        List<Page> list2 = [page1, page3]
-        List<Page> expectedResult = [page1, page2, page3]
-
-        when: "merging lists"
-        List<Page> result = controllerUnderTest.getOrderedPagesOfAllScripts([list1, list2])
-
-        then: "result should be expectedResult"
-        result == expectedResult
-    }
-
-    def "test merge two lists with different lengths"() {
-        given: "two lists that are not equal"
-        List<Page> list1 = [page1, page2]
-        List<Page> list2 = [page1, page2, page3]
-        List<Page> expectedResult = [page1, page2, page3]
-
-        when: "merging lists"
-        List<Page> result = controllerUnderTest.getOrderedPagesOfAllScripts([list1, list2])
-
-        then: "result should be expectedResult"
-        result == expectedResult
-    }
-
-    def "test merge three different lists"() {
-        given: "three lists that are not equal and have different lengths"
-        List<Page> list1 = [page1, page2, page3]
-        List<Page> list2 = [page1, page3, page3]
-        List<Page> list3 = [page1, page2, page3, page2]
-        List<Page> expectedResult = [page1, page2, page3, page3, page2]
-
-        when: "merging lists"
-        List<Page> result = controllerUnderTest.getOrderedPagesOfAllScripts([list1, list2, list3])
-
-        then: "result should be expectedResult"
-        result == expectedResult
-    }
-
-    def "convert multiple barchart aggregations to DTO"() {
-        given: "two barchart aggregations"
-        JobGroup jobGroup1 = JobGroup.build(name: "job group 1", csiConfiguration: CsiConfiguration.build())
-        JobGroup jobGroup2 = JobGroup.build(name: "job group 2", id: 2)
-        SelectedMeasurand selectedMeasurand1 = new SelectedMeasurand("DOC_COMPLETE_TIME", CachedView.CACHED)
-        SelectedMeasurand selectedMeasurand2 = new SelectedMeasurand("SPEED_INDEX", CachedView.UNCACHED)
-        List<BarchartAggregation> barchartAggregations = [
-                new BarchartAggregation(
-                        value: 1.2,
-                        valueComparative: 2.4,
-                        unit: selectedMeasurand1.measurandGroup.unit.label,
-                        measurandLabel: "de.iteratec.isr.measurand.${selectedMeasurand1.name};${selectedMeasurand1.name}",
-                        measurandName: selectedMeasurand1.name,
-                        measurandGroup: selectedMeasurand1.measurandGroup,
-                        page: page1,
-                        jobGroup: jobGroup1,
-                        aggregationValue: "foo"
-                ),
-                new BarchartAggregation(
-                        value: 3.2,
-                        valueComparative: null,
-                        unit: selectedMeasurand2.measurandGroup.unit.label,
-                        measurandLabel: "de.iteratec.isr.measurand.${selectedMeasurand2.name};${selectedMeasurand2.name}",
-                        measurandName: selectedMeasurand2.name,
-                        measurandGroup: selectedMeasurand2.measurandGroup,
-                        page: page2,
-                        jobGroup: jobGroup2,
-                        aggregationValue: "bar"
-                )
-        ]
-        controllerUnderTest.i18nService = Stub(I18nService) {
-            msg(_ as String, _ as String) >> { String key, String defaultValue -> key + ";" + defaultValue }
-        }
-
-        when: "being converted to DTO"
-        def seriesDTOs = controllerUnderTest.convertToAggregationChartSeriesDTOs(barchartAggregations)
-
-        then: "the resulting list contains DTOs for both aggregations"
-        seriesDTOs == [
-                new AggregationChartSeriesDTO(
-                        measurand: "DOC_COMPLETE_TIME",
-                        measurandLabel: "de.iteratec.isr.measurand.DOC_COMPLETE_TIME;DOC_COMPLETE_TIME",
-                        measurandGroup: "LOAD_TIMES",
-                        value: 1.2,
-                        valueComparative: 2.4,
-                        page: "page one",
-                        jobGroup: "job group 1",
-                        browser: null,
-                        unit: "ms",
-                        aggregationValue: "foo"),
-                new AggregationChartSeriesDTO(
-                        measurand: "SPEED_INDEX",
-                        measurandLabel: "de.iteratec.isr.measurand.SPEED_INDEX;SPEED_INDEX",
-                        measurandGroup: "LOAD_TIMES",
-                        value: 3.2,
-                        valueComparative: null,
-                        page: "page two",
-                        jobGroup: "job group 2",
-                        browser: null,
-                        unit: "ms",
-                        aggregationValue: "bar")
-        ]
+    void "test something"() {
+        expect:"fix me"
+            true == false
     }
 }
